Make irq_id 64 bits.
- So that DMA and other interrupts can fit without overlap.
- To simplify conversion to interrupt mask.
- To reduce variables, defines, and code.

+++ devices/common/ata/idechannel.h
@@ -95,7 +95,7 @@ private:
 
     // interrupt stuff
     InterruptCtrl*  int_ctrl = nullptr;
-    uint32_t        irq_id   = 0;
+    uint64_t        irq_id   = 0;
 };
 
 #endif // IDE_CHANNEL_H
