+++ devices/common/scsi/sc53c94.h
@@ -290,7 +290,7 @@ private:
 
     // interrupt related stuff
     InterruptCtrl* int_ctrl = nullptr;
-    uint32_t       irq_id   = 0;
+    uint64_t       irq_id   = 0;
     uint8_t        irq      = 0;
 
     // DMA related stuff
